+++ feature-renderer-card/feature-renderer-card-loader.js
@@ -1,6 +1,6 @@
 import { HAControlLoader } from "../ha-control-loader.js?v=0.6.0";
 
-const VERSION = "0.1.35";
+const VERSION = "0.1.36";
 
 const SCRIPT_NAME = "feature-renderer-card-loader.js";